filesystem copy module now uses only node
